Changed pagination count param: int -> int64

+++ httputils/json_response.go
@@ -34,7 +34,7 @@ func WriteJsonSimple(w http.ResponseWriter, statusCode int, data any) {
 // WriteJsonPaginated writes a paginated JSON response to the provided http.ResponseWriter.
 // It includes the paginated items, pagination metadata,
 // and handles potential errors during pagination or JSON encoding.
-func WriteJsonPaginated[T any](w http.ResponseWriter, r *http.Request, items []T, count int, currentPage int, limit int) {
+func WriteJsonPaginated[T any](w http.ResponseWriter, r *http.Request, items []T, count int64, currentPage int, limit int) {
 	writeJsonHeader(w, 200)
 	pagination, err := NewPagination(r, items, count, currentPage, limit)
 	if err != nil {
